Unit test failure: RegexUtilityTest::testGetAllMatches()
Closes #33
RegexUtility::getFirstNMatches() returned an empty list if the regex
didn't match the whole string as such. Fixed it by removing the check.

Another issue was that if -1 is specified as the number of matches to
be matched, it returns an empty list. Fixed that by restricting the
check to return an empty list only when N == 0.